Move llmModels off env. It's too... backend specific
