+++ logicle/app/layouts/AuthLayout.tsx
@@ -1,3 +1,5 @@
+'use server'
+
 import app from '@/lib/app'
 import { useTranslation } from 'react-i18next'
 import Image from 'next/image'
