+++ logicle/lib/chat/index.ts
@@ -21,6 +21,7 @@ import { assistantFiles } from '@/models/assistant'
 import { getBackends } from '@/models/backend'
 import { LlmModel, LlmModelCapabilities, llmModelNoCapabilities } from './models'
 import { claudeThinkingBudgetTokens } from './models/anthropic'
+import { llmModels } from '../models'
 
 export interface Usage {
   promptTokens: number
@@ -174,7 +175,7 @@ export class ChatAssistant {
     knowledge: dto.AssistantFile[] | undefined
   ) {
     this.functions = functions
-    this.llmModel = env.chat.models.find((m) => m.id == assistantParams.model)
+    this.llmModel = llmModels.find((m) => m.id == assistantParams.model)
     this.llmModelCapabilities = this.llmModel?.capabilities ?? llmModelNoCapabilities
     this.saveMessage = options.saveMessage || (async () => {})
     this.updateChatTitle = options.updateChatTitle || (async () => {})
@@ -692,7 +693,7 @@ export class ChatAssistant {
     const bestBackend = backends.reduce((maxItem, currentItem) =>
       providerScore(currentItem) > providerScore(maxItem) ? currentItem : maxItem
     )
-    const models = env.chat.models.filter((m) => m.provider == bestBackend.providerType)
+    const models = llmModels.filter((m) => m.provider == bestBackend.providerType)
     if (models.length === 0) return undefined // should never happen
     const bestModel = models.reduce((maxItem, currentItem) =>
       modelScore(currentItem.id) > modelScore(maxItem.id) ? currentItem : maxItem
