+++ logicle/lib/env.ts
@@ -1,27 +1,6 @@
-import * as fs from 'fs'
-import * as path from 'path'
-import { stockModels, LlmModel } from './chat/models'
-import { logger } from './logging'
-
 function parseOptionalInt(text?: string) {
   if (text == undefined) return undefined
   return parseInt(text)
-}
-
-const loadModels = async (dir: string) => {
-  const children = fs.readdirSync(dir).sort()
-  const readModels = async (name: string) => {
-    const childPath = path.resolve(dir, name)
-    const content = await fs.promises.readFile(childPath, 'utf-8')
-    try {
-      return (await JSON.parse(content)) as LlmModel[]
-    } catch (e) {
-      logger.error(`Invalid models file ${childPath}`, e)
-    }
-    return []
-  }
-  const fragments = await Promise.all(children.map((child) => readModels(child)))
-  return fragments.flatMap((s) => s)
 }
 
 const env = {
@@ -116,9 +95,6 @@ const env = {
       allowedFormats: process.env.CHAT_ATTACHMENTS_ALLOWED_FORMATS ?? '',
       maxImgDimPx: parseInt(process.env.CHAT_ATTACHMENTS_MAX_IMG_DIM_PX ?? '2048'),
     },
-    models: process.env.PROVISION_MODELS_PATH
-      ? await loadModels(process.env.PROVISION_MODELS_PATH)
-      : stockModels,
   },
   provision: {
     config: process.env.PROVISION_PATH,
